+++ braincube_connector/constants.py
@@ -5,6 +5,10 @@
 import os
 
 DEFAULT_CONFIG = "config.json"
+DEFAULT_VARIABLE_NAME_KEY = "standard"
+DEFAULT_VARIABLE_BCID_KEY = "bcId"
+DEFAULT_BCID_KEY = "name"
+DEFAULT_NAME_KEY = "bcId"
 DEFAULT_HOME_CONFIG = os.path.join(os.path.expanduser("~"), ".braincube", DEFAULT_CONFIG)
 NO_CONFIG_MSG = "The client needs a configuration file."
 DEFAULT_PAGE_SIZE = 150