+++ tests/test_parameters.py
@@ -14,6 +14,7 @@ def test_set_parameters(mocker):
 def test_get_parameters(mocker):
     mocker.patch.dict("braincube_connector.parameters._default_parameters", {"page_size": 8})
     mocker.patch.dict("braincube_connector.instances.instances", {"parameter_set": {}})
+    assert parameters.get_parameter("Nonexistent") is None
     assert parameters.get_parameter("page_size") == 8
     parameters.set_parameter({"page_size": 5})
     assert parameters.get_parameter("page_size") == 5