+++ braincube_connector/memory_base/nested_resources/variable.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 
+import warnings
+
 from braincube_connector.memory_base.nested_resources import mb_child
 
 
@@ -25,3 +27,12 @@ class VariableDescription(mb_child.MbChild):
             An extended variable id 'long_mb_id/dvar_id'.
         """
         return "mb{mb}/d{var}".format(mb=self._memory_base.get_bcid(), var=self._bcid)
+
+    def get_uuid(self):
+        """Get the long id instead of the uuid.
+
+        Returns:
+            a variable long id.
+        """
+        warnings.warn("VariableDescription do not have uuid, returning long_id instead.")
+        return self.get_long_id()