+++ tests/mock.py
@@ -66,7 +66,7 @@ def mock_request_entity(mocker, monkeypatch):
 @pytest.fixture
 def bc_obj():
     """Create a mock of the Braincube object."""
-    obj = braincube.Braincube(bcid="id123", name="bcname", metadata={"meta": 1},)
+    obj = braincube.Braincube(product_id="id123", name="bcname", metadata={"meta": 1},)
     return obj